refactor: yaml config
