--- cli/config/upgrade/v0.34.0/upgrade.go
@@ -1,2 +1,0 @@
-//nolint:revive,stylecheck
-package v0_34_0
