+++ cli/config/yamlstore/upgrades.go
@@ -1,4 +1,4 @@
-package config
+package yamlstore
 
 import (
 	"context"
