+++ cli/run_test.go
@@ -12,6 +12,7 @@ import (
 	"testing"
 
 	"github.com/neilotoole/slogt"
+	"github.com/neilotoole/sq/cli/config/yamlstore"
 
 	"github.com/stretchr/testify/require"
 
@@ -38,7 +39,7 @@ func newTestRunCtx(ctx context.Context, t testing.TB, cfgStore config.Store,
 		var cfgDir string
 		cfgDir, err = os.MkdirTemp("", "sq_test")
 		require.NoError(t, err)
-		cfgStore = &config.YAMLFileStore{Path: filepath.Join(cfgDir, "sq.yml")}
+		cfgStore = &yamlstore.YAMLFileStore{Path: filepath.Join(cfgDir, "sq.yml")}
 		cfg = config.New()
 		require.NoError(t, cfgStore.Save(ctx, cfg))
 	} else {
