+++ testh/testh.go
@@ -12,6 +12,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/neilotoole/sq/cli/config/yamlstore"
 	"github.com/neilotoole/sq/libsq/core/ioz"
 
 	"github.com/neilotoole/sq/libsq/core/lg/lga"
@@ -698,7 +699,7 @@ func mustLoadCollection(ctx context.Context, t testing.TB) *source.Collection {
 		return []byte(proj.Expand(string(data))), nil
 	}
 
-	fs := &config.YAMLFileStore{Path: proj.Rel(testsrc.PathSrcsConfig), HookLoad: hookExpand}
+	fs := &yamlstore.YAMLFileStore{Path: proj.Rel(testsrc.PathSrcsConfig), HookLoad: hookExpand}
 	cfg, err := fs.Load(ctx)
 	require.NoError(t, err)
 	require.NotNil(t, cfg)
